# backend/test/swagger/oss_api.yaml
swagger: '2.0'
info:
  title: OSS管理系统API
  description: OSS管理系统后端API接口文档
  version: "1.0.0"
  contact:
    email: admin@example.com
host: localhost:8080
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: 在请求头中添加 "Bearer {token}" 进行身份验证
...
paths:
...
  /oss/configs/{id}/default:
    put:
      tags:
        - oss-configs
      summary: 设置默认配置
      operationId: setDefaultConfig
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: 配置ID
      responses:
        200:
          description: 设置成功
          schema:
            $ref: '#/definitions/Response'
        401:
          description: 未授权
          schema:
            $ref: '#/definitions/ErrorResponse'
        403:
          description: 无权限操作
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: 配置不存在
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: 服务器错误
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  Response:
    type: object
    properties:
      code:
        type: integer
        example: 0
      message:
        type: string
        example: 操作成功
      data:
        type: object
        description: 空对象或其他信息
  ErrorResponse:
    type: object
    properties:
      code:
        type: integer
        example: 1001
      message:
        type: string
        example: 参数错误
      data:
        type: object
        description: 错误详情
...